# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Casefile API
servers:
  - url: /api
    description: Base API path

security:
  - SessionCookieAuth: []

tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: cases
    description: Forensic cases
  - name: artifacts
    description: Evidence artifacts (immutable once uploaded)
  - name: steps
    description: Agent reasoning steps
  - name: logs
    description: Raw tool execution audit trail
  - name: reports
    description: Final incident reports
  - name: integrity
    description: Evidence chain of custody and integrity proofs
  - name: agent
    description: Autonomous investigation agent (SSE streaming)
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      security: []
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      security: []
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      security: []
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      security: []
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the `returnTo` path on success (defaults to `/`) or `/api/login` on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      security: []
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      security: []
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  /cases:
    post:
      operationId: createCase
      tags: [cases]
      summary: Create a case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseRequest"
      responses:
        "201":
          description: Case created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
    get:
      operationId: listCases
      tags: [cases]
      summary: List cases
      description: Returns up to 200 cases, newest first.
      responses:
        "200":
          description: Cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Case"

  /cases/{caseId}:
    parameters:
      - $ref: "#/components/parameters/CaseId"
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a case (hydrated with artifacts, steps, logs, and report)
      responses:
        "200":
          description: Case detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteCase
      tags: [cases]
      summary: Delete a case (cascades to all artifacts, steps, logs, and reports)
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"

  /cases/{caseId}/artifacts:
    parameters:
      - $ref: "#/components/parameters/CaseId"
    post:
      operationId: createArtifact
      tags: [artifacts]
      summary: Upload an evidence artifact
      description: |
        Server computes the SHA256 hash and byte size from the submitted content.
        For `contentEncoding=text`, content is capped at 10 MB of UTF-8.
        For `contentEncoding=base64`, the decoded payload is capped at 64 MB.
        Larger requests return 413.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArtifactRequest"
      responses:
        "201":
          description: Artifact stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
    get:
      operationId: listCaseArtifacts
      tags: [artifacts]
      summary: List a case's artifacts (metadata only — no content)
      responses:
        "200":
          description: Artifacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artifact"
        "404":
          $ref: "#/components/responses/NotFound"

  /artifacts/{artifactId}:
    parameters:
      - $ref: "#/components/parameters/ArtifactId"
    get:
      operationId: getArtifact
      tags: [artifacts]
      summary: Get a single artifact with full content
      responses:
        "200":
          description: Artifact with content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactWithContent"
        "404":
          $ref: "#/components/responses/NotFound"

  /cases/{caseId}/steps:
    parameters:
      - $ref: "#/components/parameters/CaseId"
    get:
      operationId: listCaseSteps
      tags: [steps]
      summary: List the analysis steps for a case in order
      responses:
        "200":
          description: Steps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnalysisStep"
        "404":
          $ref: "#/components/responses/NotFound"

  /cases/{caseId}/logs:
    parameters:
      - $ref: "#/components/parameters/CaseId"
    get:
      operationId: listCaseLogs
      tags: [logs]
      summary: List the raw execution logs for a case
      responses:
        "200":
          description: Execution logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExecutionLog"
        "404":
          $ref: "#/components/responses/NotFound"

  /steps/{stepId}/logs:
    parameters:
      - $ref: "#/components/parameters/StepId"
    get:
      operationId: listStepLogs
      tags: [logs]
      summary: List the execution logs produced by a single analysis step
      responses:
        "200":
          description: Execution logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExecutionLog"
        "404":
          $ref: "#/components/responses/NotFound"

  /cases/{caseId}/chain-of-custody:
    parameters:
      - $ref: "#/components/parameters/CaseId"
    get:
      operationId: getCaseChainOfCustody
      tags: [integrity]
      summary: Chain-of-custody report for a case
      description: |
        Returns every artifact-touching tool invocation for the case in
        chronological order, including the SHA-256 hash that was verified at
        read time. This is the auditable record that proves which evidence
        the agent looked at and that it was not tampered with.
      responses:
        "200":
          description: Chain of custody
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChainOfCustody"
        "404":
          $ref: "#/components/responses/NotFound"

  /cases/{caseId}/investigate:
    parameters:
      - $ref: "#/components/parameters/CaseId"
    post:
      operationId: investigateCase
      tags: [agent]
      summary: Run the autonomous agent against a case (SSE stream)
      description: |
        Streams Server-Sent Events as the agent reasons, calls tools, records
        findings, and finalizes a report.

        Event types: `started`, `iteration`, `thinking`, `tool_call`,
        `tool_result`, `finding`, `finalized`, `tokens`, `error`, `done`.

        The response is a long-lived `text/event-stream` and cannot be
        modelled by a normal JSON Orval client — consume it with a raw
        `fetch` + `ReadableStream` reader.
      responses:
        "200":
          description: SSE stream of agent events
          content:
            text/event-stream:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"

  /cases/{caseId}/report:
    parameters:
      - $ref: "#/components/parameters/CaseId"
    get:
      operationId: getCaseReport
      tags: [reports]
      summary: Get the incident report for a case
      responses:
        "200":
          description: Incident report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentReport"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  parameters:
    CaseId:
      name: caseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ArtifactId:
      name: artifactId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    StepId:
      name: stepId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    PayloadTooLarge:
      description: Request body exceeded the size limit
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  securitySchemes:
    SessionCookieAuth:
      type: apiKey
      in: cookie
      name: sid
      description: HttpOnly session cookie set after OIDC login.

  schemas:
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code (snake_case)
        message:
          type: string
          description: Human-readable explanation
        details:
          type: object
          additionalProperties: true
          description: Optional structured context

    CaseStatus:
      type: string
      enum: [pending, analyzing, complete, failed]
    ArtifactKind:
      type: string
      enum: [log_file, network_capture, memory_strings, text, mcp_endpoint, disk_image]
    ContentEncoding:
      type: string
      enum: [text, base64]
      description: |
        How `content` is encoded in the request body. `text` (default) stores the
        UTF-8 bytes directly; `base64` is required for binary artifacts (e.g.
        disk_image) and is decoded server-side before hashing.
    AnalysisPhase:
      type: string
      enum: [triage, deep_analysis, synthesis, self_correction]

    CreateCaseRequest:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          minLength: 1
          maxLength: 10000

    Case:
      type: object
      required: [id, title, description, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/CaseStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CaseDetail:
      type: object
      required: [case, artifacts, steps, logs]
      properties:
        case:
          $ref: "#/components/schemas/Case"
        artifacts:
          type: array
          items:
            $ref: "#/components/schemas/Artifact"
        steps:
          type: array
          items:
            $ref: "#/components/schemas/AnalysisStep"
        logs:
          type: array
          items:
            $ref: "#/components/schemas/ExecutionLog"
        report:
          oneOf:
            - $ref: "#/components/schemas/IncidentReport"
            - type: "null"

    CreateArtifactRequest:
      type: object
      required: [kind, content]
      properties:
        kind:
          $ref: "#/components/schemas/ArtifactKind"
        filename:
          type: string
          maxLength: 255
        content:
          type: string
          description: |
            Raw evidence content. When `contentEncoding` is `text` (default) the
            UTF-8 byte length is capped at 10 MB. When `contentEncoding` is
            `base64` the decoded byte length is capped at 64 MB.
        contentEncoding:
          $ref: "#/components/schemas/ContentEncoding"

    Artifact:
      type: object
      required: [id, caseId, kind, sha256Hash, sizeBytes, createdAt]
      properties:
        id:
          type: string
          format: uuid
        caseId:
          type: string
          format: uuid
        kind:
          $ref: "#/components/schemas/ArtifactKind"
        filename:
          type: string
          nullable: true
        sha256Hash:
          type: string
          minLength: 64
          maxLength: 64
        sizeBytes:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time

    ArtifactWithContent:
      type: object
      required: [id, caseId, kind, sha256Hash, sizeBytes, createdAt, content]
      properties:
        id:
          type: string
          format: uuid
        caseId:
          type: string
          format: uuid
        kind:
          $ref: "#/components/schemas/ArtifactKind"
        filename:
          type: string
          nullable: true
        sha256Hash:
          type: string
          minLength: 64
          maxLength: 64
        sizeBytes:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        content:
          type: string

    AnalysisStep:
      type: object
      required:
        [
          id,
          caseId,
          stepNumber,
          phase,
          rationale,
          expected,
          found,
          nextStep,
          tokensUsed,
          durationMs,
          createdAt,
        ]
      properties:
        id:
          type: string
          format: uuid
        caseId:
          type: string
          format: uuid
        stepNumber:
          type: integer
          minimum: 1
        phase:
          $ref: "#/components/schemas/AnalysisPhase"
        toolUsed:
          type: string
          nullable: true
        rationale:
          type: string
        expected:
          type: string
        found:
          type: string
        nextStep:
          type: string
        tokensUsed:
          type: integer
          minimum: 0
        durationMs:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time

    ExecutionLog:
      type: object
      required:
        [
          id,
          caseId,
          toolName,
          input,
          output,
          tokensPrompt,
          tokensCompletion,
          startedAt,
          endedAt,
        ]
      properties:
        id:
          type: string
          format: uuid
        caseId:
          type: string
          format: uuid
        analysisStepId:
          type: string
          format: uuid
          nullable: true
        artifactId:
          type: string
          format: uuid
          nullable: true
        toolName:
          type: string
        input:
          description: Arbitrary JSON input passed to the tool
        output:
          description: Arbitrary JSON output returned by the tool
        tokensPrompt:
          type: integer
          minimum: 0
        tokensCompletion:
          type: integer
          minimum: 0
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        error:
          type: string
          nullable: true

    ChainOfCustodyEntry:
      type: object
      required: [executionLogId, artifactId, artifactSha256, artifactKind, artifactFilename, toolName, readAt, ok]
      properties:
        executionLogId:
          type: string
          format: uuid
        artifactId:
          type: string
          format: uuid
        artifactSha256:
          type: string
          nullable: true
          description: |
            SHA-256 hash recomputed and verified at the moment of the read.
            null when the read failed (e.g. integrity violation) — in that
            case the stored hash is intentionally NOT used as a fallback,
            because it cannot be claimed as verified.
        artifactKind:
          type: string
        artifactFilename:
          type: string
          nullable: true
        toolName:
          type: string
        analysisStepId:
          type: string
          format: uuid
          nullable: true
        readAt:
          type: string
          format: date-time
        ok:
          type: boolean
        error:
          type: string
          nullable: true

    ChainOfCustody:
      type: object
      required: [caseId, artifactCount, readCount, entries]
      properties:
        caseId:
          type: string
          format: uuid
        artifactCount:
          type: integer
          minimum: 0
          description: Number of distinct artifacts that have been read at least once
        readCount:
          type: integer
          minimum: 0
        entries:
          type: array
          items:
            $ref: "#/components/schemas/ChainOfCustodyEntry"

    IncidentReport:
      type: object
      required:
        [id, caseId, summary, severity, iocs, ttps, timeline, recommendations, generatedAt]
      properties:
        id:
          type: string
          format: uuid
        caseId:
          type: string
          format: uuid
        summary:
          type: string
        severity:
          type: string
          enum: [informational, low, medium, high, critical]
        iocs:
          type: array
          items:
            type: object
            additionalProperties: true
        ttps:
          type: array
          items:
            type: object
            additionalProperties: true
        timeline:
          type: array
          items:
            type: object
            additionalProperties: true
        confidenceScore:
          type: number
          nullable: true
          minimum: 0
          maximum: 1
        recommendations:
          type: array
          items:
            type: object
            additionalProperties: true
        generatedAt:
          type: string
          format: date-time